DAY18-hackLLM/model/requirements.txt
# Core dependencies for THI Pipeline
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# NLP and ML models
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
spacy>=3.7.0

# Text processing and utilities
nltk>=3.8
PyYAML>=6.0
dateparser>=1.1.0

# HTTP and API utilities
requests>=2.31.0
httpx>=0.25.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: For better performance
accelerate>=0.24.0
bitsandbytes>=0.41.0
